Add Donation Record Codes
Add Donation Record Form (previously in html format) - changed the html file type to php because sessions can only work with php files. Also added the logic which is the php code that processes data collected and links to database. I have tested the population of data into the database table and it works. In addition, added a draft home page ( with the text Home page coming soon) - which donors will be redirected to after form has been submitted.